+++ R/utils.r
@@ -189,3 +189,8 @@ menu <- function(...) {
 file.info <- function(...) {
   base::file.info(...)
 }
+
+escape_special_regex <- function(x) {
+  chars <- c("*", ".", "?", "^", "+", "$", "|", "(", ")", "[", "]", "{", "}", "\\")
+  gsub(paste0("([\\", paste0(collapse = "\\", chars), "])"), "\\\\\\1", x, perl = TRUE)
+}